only move files that actually exist, so move_file can be used after regenerating just some images of a post
--HG--
branch : moe
extra : convert_revision : svn%3A2d28d66d-8d94-df11-8c86-00306ef368cb/trunk/moe%40341

+++ app/models/post/image_store/local_flat_with_amazon_s3_backup.rb
@@ -1,10 +1,12 @@
 module PostImageStoreMethods
   module LocalFlatWithAmazonS3Backup
     def move_file
-      FileUtils.mv(tempfile_path, file_path)
-      FileUtils.chmod(0664, file_path)
+      if File.exists?(tempfile_path)
+        FileUtils.mv(tempfile_path, file_path)
+        FileUtils.chmod(0664, file_path)
+      end
 
-      if image?
+      if File.exists?(tempfile_preview_path)
         FileUtils.mv(tempfile_preview_path, preview_path)
         FileUtils.chmod(0664, preview_path)
       end
