+++ app/models/post/image_store/local_hierarchy.rb
@@ -66,11 +66,13 @@ module PostImageStoreMethods
     end
 
     def move_file
-      FileUtils.mkdir_p(File.dirname(file_path), :mode => 0775)
-      FileUtils.mv(tempfile_path, file_path)
-      FileUtils.chmod(0664, file_path)
+      if File.exists?(tempfile_path)
+        FileUtils.mkdir_p(File.dirname(file_path), :mode => 0775)
+        FileUtils.mv(tempfile_path, file_path)
+        FileUtils.chmod(0664, file_path)
+      end
 
-      if image?
+      if File.exists?(tempfile_preview_path)
         FileUtils.mkdir_p(File.dirname(preview_path), :mode => 0775)
         FileUtils.mv(tempfile_preview_path, preview_path)
         FileUtils.chmod(0664, preview_path)
